% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heapsort.R
\name{build_heap}
\alias{build_heap}
\title{Build heap matrix}
\usage{
build_heap(points, comp_fun)
}
\arguments{
\item{points}{A n by 2 matrix where each row represents a point.}

\item{comp_fun}{A compare function which states compare criterion in heap.}
}
\value{
A matrix of points with heap property.
}
\description{
Build heap matrix
}
\examples{
x<-c(2, 9, 7, 6, 5, 8)
y<-c(1, 2, 3, 4, 5, 6)
m<-matrix(c(x, y), nrow = length(x), ncol = 2)
compare_function1 <- function(p1, p2) {
  comparison_criterion <- p1[1] > p2[1]
}
build_heap(m, compare_function1)
}
