% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GridOrdering.R
\name{grid_box_ordering}
\alias{grid_box_ordering}
\title{Order points in grid box ordering algorithm}
\usage{
grid_box_ordering(points, comp_fun, grid_x_dim, grid_y_dim)
}
\arguments{
\item{points}{A n by 2 matrix where each row represents a point.}

\item{comp_fun}{A compare function which states compare criterion in heap.}

\item{grid_x_dim}{The number of grid boxes along the x-axis.}

\item{grid_y_dim}{The number of grid boxes along the y-axis.}
}
\value{
A matrix of ordered points.
}
\description{
Order points in grid box ordering algorithm
}
\examples{
set.seed(42)
points <- matrix(runif(200, 0, 1), ncol = 2)
grid_x_dim <- 2  # number of grid boxes in x direction
grid_y_dim <- 2  # number of grid boxes in y direction
compare_mmd <- function(p1, p2) {
comparison_criterion <- p1[3] > p2[3]}
grid_box_ordering(points, compare_mmd, grid_x_dim, grid_y_dim)
}
