% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heapsort.R
\name{heapify}
\alias{heapify}
\title{Check point on each index to fit heap property}
\usage{
heapify(points, index, comp_fun)
}
\arguments{
\item{points}{A n by 2 matrix where each row represents a point.}

\item{index}{Point index.}

\item{comp_fun}{A compare function which states compare criterion in heap.}
}
\value{
A matrix of points.
}
\description{
Check point on each index to fit heap property
}
\examples{
x<-c(2, 9, 7, 6, 5, 8)
y<-c(1, 2, 3, 4, 5, 6)
m<-matrix(c(x, y), nrow = length(x), ncol = 2)
compare_function1 <- function(p1, p2) {
  comparison_criterion <- p1[1] > p2[1]
}
heapify(m, 2, compare_function1)
}
