% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heapsort_MMD.R
\name{max_heapify}
\alias{max_heapify}
\title{Use maximum minimum distance to order points}
\usage{
max_heapify(Matrix, index, comp_fun)
}
\arguments{
\item{Matrix}{A n by 2 matrix where each row represents a point.}

\item{comp_fun}{A compare function which states compare criterion in heap.}
}
\value{
A matrix of ordered points.
}
\description{
Use maximum minimum distance to order points
}
\examples{
x <- c(1, 3, 0, 2, 2)
y <- c(5, 3, 1, 2, 0)
m <- matrix(c(x, y), nrow = length(x), ncol = 2)
heap_mmd_ver2(m, compare_mmd)
}
